requirements.txt
# API framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data collection
psutil>=5.9.8
pynput>=1.7.6

# Machine learning
scikit-learn>=1.5.0
joblib>=1.4.2

# Deep learning (CPU version — no GPU needed)
tensorflow-cpu>=2.17.0

# Data processing
numpy>=1.26.4
pandas>=2.2.2

# Utilities
python-dotenv>=1.0.1
pymongo>=4.6.0